+++ Sources/Internal/Extensions/String.swift
@@ -91,4 +91,8 @@ public extension String {
         }
         return self
     }
+
+    func orNilIfBlank() -> String? {
+        trimmingCharacters(in: .whitespacesAndNewlines).isEmpty ? nil : self
+    }
 }